Update branch reference and package details in GitHub workflows
The build workflows have been updated to change the target branch from 'main' to 'devtekve'. In addition, the packaging information has been updated to include a custom package id and description to highlight that this is a fork for individual development and comes with no guarantees. Changes were made in both build.yml and build_creds_aws_sdk.yml files.